major changes

+++ io.h
@@ -2,7 +2,7 @@
 #define IO_H
 
 typedef unsigned char  u8;
-typedef unsigned short u16;
+typedef unsigned int  u16;
 typedef unsigned long  u32;
 
 #define  GREEN  10         // color byte for putc()
